python3Packages.torch_2_6: use ROCm 6.2.4
We have to pass the final `rocmPackages` for downstream users to be
able to override it. I am not 100% sure why final was not used before,
but it seems like `callPackage` in the Python package extensions holds
on to the current overlay (given that the overrides of mkl etc. do work).