Merge with kernel builder packages (#60)
* python3Packages.torch: 2.6.0 -> 2.7.0

* pythonPackages.torch_2_6: init at 2.6.0

* Add some more information about this flake

* Merge in ROCm overlay

* nixpkgs with CUDA 12.9

* Merge more changes from kernel-builder/rocm-nix

* toml2cmake -> build2toml: 0.2.1

* Update hub kernels

* Don't build Torch 2.6 (for now?)

* fetchKernel: support passing rev xor version

* Use punica kernels from the hub

* Use EETQ kernels from the hub

* Bump nixpkgs to latest CUDA 12.9 branch revision

* python3Packages.torch_2_6: use ROCm 6.2.4

We have to pass the final `rocmPackages` for downstream users to be
able to override it. I am not 100% sure why final was not used before,
but it seems like `callPackage` in the Python package extensions holds
on to the current overlay (given that the overrides of mkl etc. do work).

* nix fmt

* fmt: use nixfmt-tree

* fetchKernel: cleanup

* Remove .orig file